memoized components in new files
